+++ tools/perf/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @veris/perf
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - veris-cli@1.2.1
+
 ## 0.1.2
 
 ### Patch Changes